depth-mapper/hugging-face-CPU-api/requirements.txt
torch
torchvision
transformers
Pillow
numpy
